requirements.txt
python-docx
PyPDF2
sentence-transformers
faiss-cpu
streamlit
pandas
fpdf